change of style style of .h and .cpp files changed CMakeLists.txt changed accordingly temporary addition of some 3rd party files
