+++ src/ClipBoard.cpp
@@ -1,0 +1,7 @@
+#include "ClipBoard.h"
+
+using namespace std;
+
+void ClipBoard::Copy(string &Str){
+    //clip::set_text("salam");
+}
