+++ src/coreinc.h
@@ -1,0 +1,14 @@
+#pragma once
+#include <ncurses.h>
+#include <cstring>
+#include <cstdio>
+#include <fstream>
+#include <iostream>
+#include <span>
+#include <map>
+#include <string>
+#include <dlfcn.h>
+#include <chrono>
+#include <tree_sitter/api.h>
+#include <nlohmann/json.hpp>
+
